♻️ refactor: time 영화/극장 비교 컴포넌트 이름 변경 및 분리
